refactor: split api.ts god object into domain-specific modules (Related to #646)
api.ts (1,236 lines, 88 exports) → 4 focused modules:

- learningApi.ts  — comprehension chat, reading eval, AI analysis,
                    dialogue history, listening, exit ticket, dictionary
- progressApi.ts  — error patterns, vocab recs, student progress,
                    story slugs, story recommendations, repeated errors
- parentApi.ts    — parent invite codes, linked children, child dashboard
- gamificationApi.ts — added get* aliases (getGamificationSummary, etc.)

api.ts trimmed to 183 lines: SessionExpiredError, session localStorage
helpers (saveActiveSession/loadActiveSession/clearActiveSession),
fetchStories, fetchStory.

Updated 29 consumer files across components, pages, and layouts.
Build: zero TypeScript errors. Tests: 73/73 pass.
